Fixed a small code change introduced by Code Formatting.

+++ src/main/java/com/velox/sloan/cmo/workflows/digitalpcr/DigitalPcrResultsParser.java
@@ -298,8 +298,8 @@ public class DigitalPcrResultsParser extends DefaultGenericPlugin {
             analyzedData.put("Channel1PosChannel2Pos", getSum(groupedData.get(key), "Channel1PosChannel2Pos"));
             analyzedData.put("Channel1PosChannel2Neg", getSum(groupedData.get(key), "Channel1PosChannel2Neg"));
             analyzedData.put("Channel1NegChannel2Pos", getSum(groupedData.get(key), "Channel1NegChannel2Pos"));
-            Integer dropletCountMutation = analyzedData.get("Channel1PosChannel2Pos") + analyzedData.get("Channel1PosChannel2Neg");
-            Integer dropletCountWildType = analyzedData.get("Channel1PosChannel2Pos") + analyzedData.get("Channel1NegChannel2Pos");
+            Integer dropletCountMutation = (Integer) analyzedData.get("Channel1PosChannel2Pos") + (Integer) analyzedData.get("Channel1PosChannel2Neg");
+            Integer dropletCountWildType = (Integer) analyzedData.get("Channel1PosChannel2Pos") + (Integer) analyzedData.get("Channel1NegChannel2Pos");
             Double totalDnaDetected = calculateTotalDnaDetected((Double) analyzedData.get("ConcentrationMutation"), (Double) analyzedData.get("ConcentrationWildType"));
             analyzedData.put("DropletCountMutation", dropletCountMutation);
             analyzedData.put("DropletCountWildType", dropletCountWildType);
